This commit was manufactured by cvs2svn to create branch 'struct-reorg-branch'.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/branches/struct-reorg-branch@87007 138bc75d-0d04-0410-961f-82ee72b054a4

+++ libbanshee/engine/stamp.h
@@ -36,7 +36,7 @@
 
 EXTERN_C_BEGIN
 
-typedef int stamp;
+typedef long stamp;
 
 stamp stamp_fresh(void);
 stamp stamp_fresh_small(void);
